+++ apps/tetris-classic/src/pages/[locale]/puzzle/index.tsx
@@ -3,7 +3,7 @@ import { getStaticPaths, makeStaticProps } from 'features/locale/localeStatic';
 import Page from 'controls/page/Page';
 import Puzzle from 'components/puzzle/Puzzle';
 
-import { client } from 'utils/utils';
+import { client } from '@muravjev/utils-core';
 console.log(client, 'compile puzzle');
 
 const getStaticProps = makeStaticProps(['common', 'page-puzzle', 'game']);