+++ packages/utils/redux-subslice/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @muravjev/utils-redux-subslice
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies \[8a568f1]
+  - @muravjev/utils-core-ts@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes